+++ CHANGELOG.md
@@ -14,9 +14,10 @@ and this project adheres to [0-based versioning](https://0ver.org/).
 - Upgrade libuv to v1.51.0
 - Move active object main state machine init transition to task context before global task lock
 - Rename:
-  - am_task_wait_all() to am_task_startup_gate_wait()
-  - am_task_lock_all() to am_task_startup_gate_close()
-  - am_task_unlock_all() to am_task_startup_gate_open()
+  - `am_task_wait_all()` to `am_task_startup_gate_wait()`
+  - `am_task_lock_all()` to `am_task_startup_gate_close()`
+  - `am_task_unlock_all()` to `am_task_startup_gate_open()`
+- Rework `am_ao_run_all()` in AO cooperative port
 
 ### Fixed